requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_rds
aws_cdk.aws_directoryservice
simplejson
aws_cdk.aws_fsx